Made StringTrieNode a class and added functions
hasSuffixNode takes an input character and returns true if that character is a valid suffix character (e.g. the character is a key value in m_paths)

getSuffixNode takes an input character and returns a pointer to the suffix node with that character. The function returns a null pointer if the suffix node doesn't exist