Fixed for loop in removeSubTrie
Range based for loop in removeSubTrie was iterating through const pairs
of map. Recursive calls to removeSubTrie would then try to delete the
second element of the const pair and causing a segfault. Pairs in for
loop are no longer const, but the first element of the pair (the char
key) must be const to compile properly.